application auth completed

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -32,7 +32,7 @@ public class JournalEntryService {
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
 //            user.setUserName(null);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }catch (Exception e){
             System.out.println(e);
             throw new RuntimeException("An error occurred while saving the entry. ",e);
@@ -55,7 +55,23 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id){
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed=false;
+        try{
+           User user= userService.findByUserName(userName);
+            removed= user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+           if(removed){
+               userService.saveUser(user);
+               journalEntryRepository.deleteById(id);
+           }
+
+        }catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleing the entry. ",e);
+        }
+        return removed;
     }
+
+
 }
